% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/innbAldKj2014_2018.R
\docType{data}
\name{innbAldKj2014_18}
\alias{innbAldKj2014_18}
\title{Norsk befolkningsdata om demografi og tilhørighet i spesialisthelsesystemet}
\format{Dataramme med 505 620 observasjoner og 14 variabler
\describe{
  \item{Aar}{}
  \item{ErMann}{}
  \item{Alder}{}
  \item{KomNr}{}
  \item{KomNrTxt}{}
  \item{Fylke}{}
  \item{FylkeNr}{}
  \item{BoHF}{}
  \item{BoHFnr}{}
  \item{BoRHF}{}
  \item{BoRHFnr}{}
  \item{Bydel}{}
  \item{BydelNr}{}
  \item{Innbyggere}{}
}}
\source{
\url{}
}
\usage{
innbAldKj2014_18
}
\description{
Et datasett med norsk befolkningsdata fra og med 2014 til og med 2018 om
demografi (alder, kjønn, (bydel), kommune, fylke) og tilhørighet i
spesialisthelsetjenesten (helseforetak, regionalt helseforetak)
}
\keyword{datasets}
